+++ client/src/components/admin/Doctor/Doctor.module.css
@@ -33,10 +33,10 @@
 
 .doctorUpdate{
     flex: 2;
-    padding: 20px;
+    /* padding: 20px;
     -webkit-box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8); 
     box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8);
-    margin-left: 20px;
+    margin-left: 20px; */
 }
 
 .doctorShowTop{
@@ -89,7 +89,7 @@
     font-size: 22px !important;
 }
 
-.doctorUpdateForm{
+/* .doctorUpdateForm{
     display: flex;
     justify-content: space-between;
     margin-top: 20px;
@@ -154,4 +154,4 @@
     background-color: teal;
     color: white;
     font-weight: 600;
-}
+} */
