+++ client/src/components/admin/Doctor/DoctorEdit.module.css
@@ -1,0 +1,73 @@
+.doctorUpdate{
+    padding: 20px;
+    -webkit-box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8); 
+    box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8);
+    margin-left: 20px;
+}
+
+.doctorUpdateTitle{
+    font-size: 25px;
+    font-weight: 600;
+    margin-left: 10px;
+}
+
+.doctorUpdateForm{
+    display: flex;
+    justify-content: space-between;
+    margin-top: 20px;
+    margin-left: 10px;
+}
+
+.doctorUpdateItem{
+    display: flex;
+    flex-direction: column;
+    margin-top: 5px;
+}
+
+.doctorUpdateItem>label{
+    margin-bottom: 5px;
+    font-weight: 500;
+}
+
+.doctorUpdateInput{
+    border: none;
+    width: 250px;
+    height: 17px;
+    border-bottom: 1px solid grey;
+}
+
+.doctorUpdateRight{
+    display: flex;
+    flex-direction: column;
+    justify-content:space-evenly;
+    align-items: center;
+}
+
+.doctorUpdateUpload{
+    display: flex;
+    align-items: center;
+}
+
+.doctorUpdateImg{
+    width: 200px;
+    height: 200px;
+    border-radius: 20px;
+    object-fit: cover;
+    margin-right: 20px;
+}
+
+/* .doctorUpdateIcon{
+    cursor: pointer;
+    font-size: 35px !important;
+} */
+
+.doctorUpdateBotton{
+    width: 150px;
+    border: none;
+    border-radius: 5px;
+    padding: 10px;
+    cursor: pointer;
+    background-color: teal;
+    color: white;
+    font-weight: 600;
+}
