+++ client/src/components/admin/User/User.module.css
@@ -1,4 +1,5 @@
-.User{
+.Patient{
+    /* display: flex; */
     flex: 4;
     padding: 0px 20px;
 }
@@ -33,10 +34,10 @@
 
 .userUpdate{
     flex: 2;
-    padding: 20px;
+    /* padding: 20px;
     -webkit-box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8); 
     box-shadow: 0px 0px 15px 0px rgba(0,0,0,0.8);
-    margin-left: 20px;
+    margin-left: 20px; */
 }
 
 .userShowTop{
@@ -89,7 +90,7 @@
     font-size: 22px !important;
 }
 
-.userUpdateForm{
+/* .userUpdateForm{
     display: flex;
     justify-content: space-between;
     margin-top: 20px;
@@ -154,4 +155,4 @@
     background-color: teal;
     color: white;
     font-weight: 600;
-}
+} */
